feat: extractImageIdsFromDelta 함수 분리
