+++ src/features/write/lib/index.ts
@@ -1,0 +1,3 @@
+import { extractImageIdsFromDelta } from './extractImageIdsFromDelta';
+
+export { extractImageIdsFromDelta };
